Split types and base client into separate files. Convert tests to typescript.
