+++ package.json
@@ -22,10 +22,10 @@
     "test-types": "node test/typescript/index.js && tsc --esModuleInterop true --noEmit true --strictNullChecks true --noImplicitAny true --strict true test/typescript/*.ts",
     "eslint": "eslint '**/*.{js,md,ts}' --max-warnings 0 --ignore-path ./.eslintignore",
     "eslint-fix": "npx eslint --fix '**/*.{js,md,ts}' --max-warnings 0 --ignore-path ./.eslintignore",
-    "test-unit": "NODE_ENV=test mocha --exit --bail --timeout 700000 --require ./babel-register test/unit/*.js",
-    "test-e2e": "NODE_ENV=test mocha --exit --bail --timeout 700000 --require ./babel-register test/e2e/*.js",
+    "test-unit": "NODE_ENV=test mocha --exit --bail --timeout 700000 --require ./babel-register test/unit/*.{js,ts}",
+    "test-e2e": "NODE_ENV=test mocha --exit --bail --timeout 700000 --require ./babel-register test/e2e/*.{js,ts}",
     "test-coverage": "nyc yarn test-unit",
-    "testwatch": "NODE_ENV=test nodemon ./node_modules/.bin/mocha --timeout 20000 --require test-entry.js test/unit/*.js",
+    "testwatch": "NODE_ENV=test nodemon ./node_modules/.bin/mocha --timeout 20000 --require test-entry.js test/unit/*.{js,ts}",
     "lint": "yarn run prettier && yarn run eslint",
     "lint-fix": "yarn run prettier-fix && yarn run eslint-fix",
     "prepare": "yarn run build",
