+++ src/client.ts
@@ -1,4 +1,3 @@
-import axios, { AxiosError, AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
 import {
   APIResponse,
   Asset,
@@ -6,11 +5,9 @@ import {
   Collection,
   TransferParams,
   UserParams,
-  OriginalOptions,
   UidResponse,
   Transfer,
   User,
-  Environment,
   AssetParams,
   EditAssetParams,
   Deposit,
@@ -20,128 +17,11 @@ import {
   ClaimParams,
   Claim,
   Reward,
+  Balance,
 } from './types';
-import { APIErrorResponse, isErrorResponse, throwErrorFromResponse } from './error';
-import { TokenManager } from './token_manager';
+import { BaseClient } from './baseClient';
 
-const DEVELOPMENT_URL = 'https://api-dev.getoriginal.com/v1';
-const PRODUCTION_URL = 'https://api.getoriginal.com/v1';
-
-export class OriginalClient {
-  private apiKey: string;
-  private secret: string;
-  private axiosInstance: AxiosInstance;
-  private options: OriginalOptions;
-  private baseURL: string;
-  private tokenManager: TokenManager;
-
-  /**
-   * Initialize a client
-   *
-   * @param {string } [apiKey] - the api key
-   * @param {string} [secret] - the api secret
-   * @param {OriginalOptions} [options] - additional options, here you can pass the env and options to axios instance
-   * @example <caption>initialize the client</caption>
-   * new OriginalClient('api_key', 'secret')
-   */
-
-  constructor(apiKey: string, secret: string, options?: OriginalOptions) {
-    if (!apiKey || !secret) {
-      throw new Error('apiKey and secret are required');
-    }
-    this.apiKey = apiKey;
-    this.secret = secret;
-
-    this.tokenManager = new TokenManager(this.apiKey, this.secret);
-    const configOptions = options ? options : {};
-
-    this.options = {
-      withCredentials: false, // making sure cookies are not sent
-      ...configOptions,
-    };
-
-    this.axiosInstance = axios.create(this.options);
-
-    this.baseURL = this.options.baseURL || this.getEnvURL(this.options.env || Environment.Production);
-  }
-
-  private getEnvURL(env: string) {
-    if (env === Environment.Development) {
-      return DEVELOPMENT_URL;
-    } else if (env === Environment.Production) {
-      return PRODUCTION_URL;
-    } else {
-      throw new Error('Invalid environment');
-    }
-  }
-
-  private doAxiosRequest = async <T>(
-    type: string,
-    endpoint: string,
-    data?: unknown,
-    options: AxiosRequestConfig & {
-      config?: AxiosRequestConfig & { maxBodyLength?: number };
-    } = {},
-  ): Promise<T> => {
-    const url = `${this.baseURL}/${endpoint}`;
-    const token = this.tokenManager.getToken();
-    const headers = {
-      'Content-Type': 'application/json',
-      'X-API-KEY': this.apiKey,
-      Authorization: `Bearer ${token}`,
-    };
-
-    const requestConfig = {
-      ...options,
-      headers,
-    };
-    try {
-      let response: AxiosResponse<T>;
-      try {
-        switch (type) {
-          case 'get':
-            response = await this.axiosInstance.get(url, requestConfig);
-            break;
-          case 'post':
-            response = await this.axiosInstance.post(url, data, requestConfig);
-            break;
-          case 'put':
-            response = await this.axiosInstance.put(url, data, requestConfig);
-            break;
-          default:
-            throw new Error('Invalid request type');
-        }
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      } catch (e: any) {
-        return this.handleResponse(e.response);
-      }
-      return this.handleResponse(response);
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } catch (e: any /**TODO: generalize error types  */) {
-      throw e as AxiosError<APIErrorResponse>;
-    }
-  };
-
-  private _get<T>(url: string, params?: AxiosRequestConfig['params']) {
-    return this.doAxiosRequest<T>('get', url, null, { params });
-  }
-
-  private _post<T>(url: string, data?: unknown) {
-    return this.doAxiosRequest<T>('post', url, data);
-  }
-
-  private _put<T>(url: string, data?: unknown) {
-    return this.doAxiosRequest<T>('put', url, data);
-  }
-
-  private handleResponse<T>(response: AxiosResponse<T>) {
-    const data = response.data;
-    if (isErrorResponse(response)) {
-      throw throwErrorFromResponse(response);
-    }
-    return data;
-  }
-
+export class OriginalClient extends BaseClient {
   /**
    * Methods to access the API
    * [Read the Docs](https://docs.getoriginal.com/docs)
@@ -156,7 +36,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<UidResponse>>} Returns the response object with the uid of the created user.
    * Will throw an error if the user already exists with the supplied details.
    */
-  public async createUser(user?: UserParams) {
+  public async createUser(user?: UserParams): Promise<APIResponse<UidResponse>> {
     return this._post<APIResponse<UidResponse>>('user', user);
   }
 
@@ -167,7 +47,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<User>>} Returns the response object with the details of the user.
    * Will throw a 404 error if the user does not exist.
    */
-  public async getUser(uid: string) {
+  public async getUser(uid: string): Promise<APIResponse<User>> {
     return this._get<APIResponse<User>>(`user/${uid}`);
   }
 
@@ -178,21 +58,8 @@ export class OriginalClient {
    * @return {Promise<APIResponse<User | null>>} Returns the response object with the details of the user,
    * or null data if not found.
    */
-  public async getUserByEmail(email: string) {
+  public async getUserByEmail(email: string): Promise<APIResponse<User | null>> {
     return this._get<APIResponse<User | null>>('user', { email });
-  }
-
-  /**
-   * @deprecated getUserByClientId. Please use `getUserByUserExternalId` instead.
-   *
-   * getUserByClientId
-   *
-   * @param {String} clientId ClientId of the user to get
-   * @return {Promise<APIResponse<User | null>>} Returns the response object with the details of the user,
-   * or null data if not found.
-   */
-  public async getUserByClientId(clientId: string) {
-    return this._get<APIResponse<User | null>>('user', { client_id: clientId });
   }
 
   /**
@@ -202,7 +69,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<User | null>>} Returns the response object with the details of the user,
    * or null data if not found.
    */
-  public async getUserByUserExternalId(userExternalId: string) {
+  public async getUserByUserExternalId(userExternalId: string): Promise<APIResponse<User | null>> {
     return this._get<APIResponse<User | null>>('user', { user_external_id: userExternalId });
   }
 
@@ -217,7 +84,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<Collection>>} Returns the response object with details of the collection.
    * Will throw a 404 error if the collection does not exist.
    */
-  public async getCollection(uid: string) {
+  public async getCollection(uid: string): Promise<APIResponse<Collection>> {
     return this._get<APIResponse<Collection>>(`collection/${uid}`);
   }
 
@@ -232,7 +99,7 @@ export class OriginalClient {
    * Will throw an error if the asset already exists, or if not all required fields.
    * in the AssetParams are provided.
    */
-  public async createAsset(asset: AssetParams) {
+  public async createAsset(asset: AssetParams): Promise<APIResponse<UidResponse>> {
     return this._post<APIResponse<UidResponse>>('asset', asset);
   }
 
@@ -242,7 +109,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<Asset>>} Returns the response object with the details of the asset.
    * Will throw a 404 error if the asset does not exist.
    */
-  public async getAsset(uid: string) {
+  public async getAsset(uid: string): Promise<APIResponse<Asset>> {
     return this._get<APIResponse<Asset>>(`asset/${uid}`);
   }
 
@@ -251,8 +118,8 @@ export class OriginalClient {
    * @param {string} userUid uid of the owner of the assets to get.
    * @return {Promise<APIResponse<Asset[]>>} Returns the response object with a list of assets owned by the user. Returns an empty list if none found.
    */
-  public async getAssetsByUserUid(userUid: string) {
-    return this._get<APIResponse<Asset[] | null>>('asset', { user_uid: userUid });
+  public async getAssetsByUserUid(userUid: string): Promise<APIResponse<Asset[]>> {
+    return this._get<APIResponse<Asset[]>>('asset', { user_uid: userUid });
   }
 
   /**
@@ -261,7 +128,7 @@ export class OriginalClient {
    * @param {EditAssetParams} asset The details of the asset to be edited.
    * @return {Promise<APIResponse<null>>} Returns the response object with the success status of the edit.
    */
-  public async editAsset(uid: string, asset: EditAssetParams) {
+  public async editAsset(uid: string, asset: EditAssetParams): Promise<APIResponse<null>> {
     return this._put<APIResponse<null>>(`asset/${uid}`, asset);
   }
 
@@ -275,7 +142,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<UidResponse>>} Returns the response object with the uid of the created transfer.
    * Will throw an error if the transfer already exists, or if not all required fields are provided.
    */
-  public async createTransfer(transfer: TransferParams) {
+  public async createTransfer(transfer: TransferParams): Promise<APIResponse<UidResponse>> {
     return this._post<APIResponse<UidResponse>>('transfer', transfer);
   }
 
@@ -285,7 +152,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<Transfer>>} Returns the response object with the details of the transfer.
    * Will throw a 404 error if the transfer does not exist.
    */
-  public async getTransfer(uid: string) {
+  public async getTransfer(uid: string): Promise<APIResponse<Transfer>> {
     return this._get<APIResponse<Transfer>>(`transfer/${uid}`);
   }
 
@@ -294,8 +161,8 @@ export class OriginalClient {
    * @param {string} userUid userUid of asset to transfer.
    * @return {Promise<APIResponse<Transfer[]>>} Returns the response object with a list of transfers made by the user. Returns an empty list if none found.
    */
-  public async getTransfersByUserUid(userUid: string) {
-    return this._get<APIResponse<Transfer[] | null>>('transfer', { user_uid: userUid });
+  public async getTransfersByUserUid(userUid: string): Promise<APIResponse<Transfer[]>> {
+    return this._get<APIResponse<Transfer[]>>('transfer', { user_uid: userUid });
   }
 
   /**
@@ -307,7 +174,7 @@ export class OriginalClient {
    * @param {BurnParams} burn The details of the burn to be created.
    * @return {Promise<APIResponse<UidResponse>>} Returns the response object with the uid of the created burn.
    */
-  public async createBurn(burn: BurnParams) {
+  public async createBurn(burn: BurnParams): Promise<APIResponse<UidResponse>> {
     return this._post<APIResponse<UidResponse>>('burn', burn);
   }
 
@@ -317,7 +184,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<Burn>>} Returns the response object with the details of the burn.
    * Will throw a 404 error if the burn does not exist.
    */
-  public async getBurn(uid: string) {
+  public async getBurn(uid: string): Promise<APIResponse<Burn>> {
     return this._get<APIResponse<Burn>>(`burn/${uid}`);
   }
 
@@ -326,8 +193,8 @@ export class OriginalClient {
    * @param {string} userUid user_uid of asset to burn.
    * @return {Promise<APIResponse<Burn[]>>} Returns the response object with a list of burns made by the user. Returns an empty list if none found.
    */
-  public async getBurnsByUserUid(userUid: string) {
-    return this._get<APIResponse<Burn[] | null>>('burn', { user_uid: userUid });
+  public async getBurnsByUserUid(userUid: string): Promise<APIResponse<Burn[]>> {
+    return this._get<APIResponse<Burn[]>>('burn', { user_uid: userUid });
   }
 
   /**
@@ -337,7 +204,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<Deposit>>} Returns the response object with the deposit details of a user.
    * Will throw a 404 error if the user does not exist.
    */
-  public async getDeposit(userUid: string, collectionUid?: string) {
+  public async getDeposit(userUid: string, collectionUid: string): Promise<APIResponse<Deposit>> {
     return this._get<APIResponse<Deposit>>('deposit', { user_uid: userUid, collection_uid: collectionUid });
   }
 
@@ -352,7 +219,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<Reward>>} Returns the details of the reward.
    * Will throw a 404 error if the reward does not exist.
    */
-  public async getReward(uid: string) {
+  public async getReward(uid: string): Promise<APIResponse<Reward>> {
     return this._get<APIResponse<Reward>>(`reward/${uid}`);
   }
 
@@ -365,7 +232,7 @@ export class OriginalClient {
    * @param {AllocationParams} allocation The details of the allocation to be created
    * @return {Promise<APIResponse<UidResponse>>} Uid of the created allocation
    */
-  public async createAllocation(allocation: AllocationParams) {
+  public async createAllocation(allocation: AllocationParams): Promise<APIResponse<UidResponse>> {
     return this._post<APIResponse<UidResponse>>('reward/allocate', allocation);
   }
 
@@ -375,7 +242,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<Allocation>>} Returns details of the allocation
    * Will throw a 404 error if the allocation does not exist.
    */
-  public async getAllocation(uid: string) {
+  public async getAllocation(uid: string): Promise<APIResponse<Allocation>> {
     return this._get<APIResponse<Allocation>>(`reward/allocate/${uid}`);
   }
 
@@ -384,8 +251,8 @@ export class OriginalClient {
    * @param {string} userUid user_uid of allocations
    * @return {Promise<APIResponse<Allocation[]>>} Returns a list of allocations available to the user, empty if none found
    */
-  public async getAllocationsByUserUid(userUid: string) {
-    return this._get<APIResponse<Allocation[] | null>>('reward/allocate', { user_uid: userUid });
+  public async getAllocationsByUserUid(userUid: string): Promise<APIResponse<Allocation[]>> {
+    return this._get<APIResponse<Allocation[]>>('reward/allocate', { user_uid: userUid });
   }
 
   /**
@@ -397,7 +264,7 @@ export class OriginalClient {
    * @param {ClaimParams} claim The details of the claim to be created
    * @return {Promise<APIResponse<UidResponse>>} Uid of the created allocation
    */
-  public async createClaim(claim: ClaimParams) {
+  public async createClaim(claim: ClaimParams): Promise<APIResponse<UidResponse>> {
     return this._post<APIResponse<UidResponse>>('reward/claim', claim);
   }
 
@@ -407,7 +274,7 @@ export class OriginalClient {
    * @return {Promise<APIResponse<Claim>>} Returns details of the allocation
    * Will throw a 404 error if the allocation does not exist.
    */
-  public async getClaim(uid: string) {
+  public async getClaim(uid: string): Promise<APIResponse<Claim>> {
     return this._get<APIResponse<Claim>>(`reward/claim/${uid}`);
   }
 
@@ -416,12 +283,17 @@ export class OriginalClient {
    * @param {string} userUid user_uid of claims
    * @return {Promise<APIResponse<Claim[]>>} Returns a list of claims available to the user, empty if none found
    */
-  public async getClaimsByUserUid(userUid: string) {
-    return this._get<APIResponse<Claim[] | null>>('reward/claim', { user_uid: userUid });
+  public async getClaimsByUserUid(userUid: string): Promise<APIResponse<Claim[]>> {
+    return this._get<APIResponse<Claim[]>>('reward/claim', { user_uid: userUid });
+  }
+
+  /**
+   * balance
+   * @param {string} rewardUid reward_uid of the reward to get balance for
+   * @param {string} userUid user_uid of the user to get balance for
+   * @return {Promise<APIResponse<number>>} Returns the balance of the user for the reward
+   */
+  public async getBalance(rewardUid: string, userUid: string): Promise<APIResponse<Balance>> {
+    return this._get<APIResponse<Balance>>('reward/balance', { reward_uid: rewardUid, user_uid: userUid });
   }
 }
-
-/**
- * @deprecated use OriginalClient instead
- */
-export { OriginalClient as Original };
