+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './client';
 export * from './types';
-export * from './token_manager';
+export * from './tokenManager';
 export * from './error';
