+++ src/tokenManager.ts
