+++ src/types/apiResponse.ts
@@ -1,0 +1,1 @@
+export type APIResponse<T> = { data: T; success: boolean };
