+++ src/types/balance.ts
@@ -1,0 +1,5 @@
+export type Balance = {
+  amount: number;
+  reward_uid: string;
+  user_uid: string;
+};
