+++ src/types/burn.ts
@@ -1,0 +1,12 @@
+export type BurnParams = {
+  asset_uid: string;
+  from_user_uid: string;
+};
+
+export type Burn = {
+  asset_uid: string;
+  created_at: string;
+  from_user_uid: string;
+  status: string;
+  uid: string;
+};
