+++ src/types/collection.ts
@@ -1,0 +1,16 @@
+export type Collection = {
+  chain: string;
+  chain_id: number;
+  contract_address: string;
+  created_at: string;
+  description: string;
+  editable_assets: boolean;
+  explorer_url: string;
+  external_url: string;
+  name: string;
+  network: string;
+  status: string;
+  symbol: string;
+  type: string;
+  uid: string;
+};
