+++ src/types/deposit.ts
@@ -1,0 +1,6 @@
+export type Deposit = {
+  chain_id: number;
+  network: string;
+  qr_code_data: string;
+  wallet_address: string;
+};
