+++ src/types/index.ts
@@ -1,0 +1,13 @@
+export * from './allocation';
+export * from './apiResponse';
+export * from './asset';
+export * from './balance';
+export * from './burn';
+export * from './claim';
+export * from './collection';
+export * from './deposit';
+export * from './edit';
+export * from './reward';
+export * from './transfer';
+export * from './uid';
+export * from './user';
