+++ src/types/reward.ts
@@ -1,0 +1,15 @@
+export type Reward = {
+  chain: string;
+  contract_address: string;
+  created_at: string;
+  description: string;
+  explorer_url: string;
+  name: string;
+  network: string;
+  status: string;
+  token_name: string;
+  token_type: string;
+  uid: string;
+  withdraw_receiver: string;
+  chain_id?: number;
+};
