+++ src/types/uid.ts
@@ -1,0 +1,1 @@
+export type UidResponse = { uid: string };
